Fixed PermanentDeath Setting Not Remembered
Previously when changing game mode to survival it would reset permanent
death to On, the function that does this is also called during load to
ensure check box is disabled if the game mode is creative, but if it was
survival, it would ensure checkbox is enabled and checked.
Now when its survival game mode, it only turns on/checks the permanent
death checkbox if on/true in settings. If the game type is creative
checkbox still defaults of off/unchecked.