Comments, new function + index checks
Commented all functions for in game interface.
Added GetQueueItemAmount function to return total amount of a particular
item in queue.
Altered index check for m_queue to use better method.
Added index check on blueprint.results just in case there is a blueprint
with no results - not sure if this could actually happen, perhaps with
modded blueprints, so best to check.
Removed unnecessary use of local variable in GetQueueItems.