Improve turret aim.
This uses a formula for aiming turrets based on more straightforward math. It gives the same result (without rounding) as the existing formula, but uses fewer floating-point operations. It is more accurate due to reduced round-off error and using double for all operations.

Also, prioritize targets based on the distance they will be in the near future, rather than the distance they are right now. This makes turrets much better at targeting meteors that are likely to hit.