Add AlphaSaturation of 1 to materials with no AlphaSaturation tag.
Added "<AlphaSaturation>1</AlphaSaturation>" to every material with no <AlphaSaturation> tag.

This change fixes a bug when running the game under WINE which causes light sources, such as thrusters and the sun, to have a solid black box around them. Even though WINE is not officially supported, the fix is so easy and doesn't affect Windows players, so there isn't a reason not to add it.