Fix characters dying suddenly when exiting cockpit.
This seems to fix the issue with characters dying instantly upon exiting a cockpit, especially while travelling at high speed, or in creative mode suddenly flying off uncontrollably. It is motivated by the observation that removing and putting back on your helmet will restore control. To be quite honest, I am not entirely sure why this works, but it seems to fix the problem.