Adds start/end events in Mysession to help profile component gamelogic
There are some blocks that are really adding a lot of gamelogic in
servers.  Finding them is very hard since there's no real way to
profile, and we need to do it on a live server.  These events are raised
before and after a component is processed so that profiling can take
place.  By default does nothing, so it should add no overhead, but gives
the ability to do custom profiling of gamelogic.