Hostile ship configuration code
Allows a group faction ID to be generated for each spawn group, which is
then assigned to each ship within that group.

Basic configuration of hostile ships, to ensure that they are as visible
and dangerous as possible in their native state.

Hostile Encounters set to default advanced settings world option.