Added 'Max # Ships in Spawn Group' option
Added a slider to the new GUI that allows you to set the maximum number
of ships in a spawn group, ranging from 1 to 10.  The default value 2,
will use the original selection functionality.  Values higher than 2
will use a modified version, which generates a random # ships up to the
maximum specified.