Bug fixes, performance improvements and refactoring.
Wow, must have been half asleep when I made the prefabManager changes.
Have cleared up a few bugs that caused the behaviour of the code to
differ significantly from the description of the change.  Have also
refactored the code to make it a little easier to follow and reduced the
impact on performance of the changes that I made earlier.