Added ship profiler and controls
Added a fast .sbc xml reader to obtain ship stats and vital info needed
by the configuration screen.
Added first implementation of the ship selector to the configuration
screen.
NOTE: Had to add 'MyPrefabProfileDefinition' POCO to hold ship stats.
Wasn't certain where to put it, so went for the same folder as the
DefinitionManager.  May need to move it to a better place later.