Damage ownership and damage events
This is a large change that modifies IMyDestroyableObject, but
effectively doesn't really do much the main game itself (but should be
used for kill notifications!).  All damage is passed around with
entities that are doing the damage.  This is then used in 3 new events
added by this PR:

a)
OnDestroyed(float damage, MyDamageType damageType, long attackerId)
damage - Amount of damage
damageType - Type of damage taken
attackerId - EntityID of attacker.  Is not parent, but actual direct
entity (for example a turret).

This event is triggered when an IMyDestroyableObject is destroyed, with
the amount of damage, damage type and entityId of the attacker who
destroyed this object.  Can obviously be used to track players killing
things or other players.

b)
float OnBeforeDamageApplied(float damage, MyDamageType damageType, long
attackerId)
damage - The amount of damage being applied
damageType - The type of damage being applied
attackerId - EntityId of the object doing damage

returns - Modified damage

This event is raised before damage is applied to a block.  This allows a
mod to modify damage before it's applied.  This sets up a situation
where a mod can create damage mitigation on certain types of damage.
For example you can mitigate some damage on reinforced blocks that are
protected against projectiles, or against explosions, etc.  This allow
allows mod to create indestructible blocks by return damage of 0.

c)
OnAfterDamageApplied(float damage, MyDamageType damageType, long
attackerId)
damage - Amount of damage applied
damageType - Type of damage applied
attackerId - EntityId of the object doing damage

This event is raised after damage is applied.  This would allow a mod to
create blocks that are "unstable" to certain types of damage.  For
example reactors that explode with even a small type of explosive
damage, etc.  This also allows mods to track damage easier in general.