Moved ship selector table to it's own GUI
Moved the encounter ship selector table to a new GUI screen.  This is
needed because it looks stupid hanging off the bottom of the encounters
screen and using the middle mouse to scroll the table would scroll the
entire screen when you got to the top or bottom.  It also means that I
can keep the focus of the encounters configuration screen on settings
that people will want to view and set regularly while allowing them to
retain the element of surprise if they don't want to see the details of
the ships that will appear.