Added control validation to damaged and antenna on controls
Damaged % control now sets Severity to 'No Damage' and checks the
'Applied Globally' checkbox.  This is just to make it clearer to the
player what the state of the game will be when no damaged ships are
present.

Same goes for the Antenna % and the Antenna maxed option.  It doesn't
really make sense to have no ships with active antenna, but still max
them out.  So 'the maxed out range' checkbox is unchecked and disabled,
when set the slider is set to 0%.