Fix item transfer within same inventory
Problem: When an item in an inventory was moved to a different slot in
the same inventory, it wouldn't change position when the volume of the
items being moved was greater than the remaining empty volume.

Affects:  All block inventories.

Cause:  The items were being added to the inventory before being removed
from it, without checking whether it was the same inventory, so the new
volume calculated had the potential to exceed the available space in the
container, even though it had no more items in it than it began with.
The process makes sense when you're moving items to another container,
but not to the same.

Resolution: Added a check to see if the source and destination
containers are the same and if they are then the items are removed first
and then added in the new position selected, otherwise the original
transfer method is used.