Fix formatting issue with block inventories
When inventories have 7 items in them, they automatically generate a new
line of controls to allow the player to move more items into the
inventory.   Everything in the inventory panel below that block, will
then be shuffled down to make space for the new controls.

Where refinery output's and assembler input's are present, this can lead
to horrible issues when trying to use the inventories below them, as
both frequently have very small quantities of an ore or ingot in the 7th
item, which is then used up almost immediately, causing the inventory
rows to grow and shrink repeatedly and rapidly over several seconds.

All inventories below that become incredibly hard to use as
(particularly with multiple refineries etc.) they jump up and down the
screen over large distances almost at random and sometimes several times
a second.

This fixes that problem, by only removing the additional row created
once the 6th item, rather than the 7th, has been removed.  A check is
made to ensure that if a 7th item has just been removed (inferred by the
number of rows present previously as opposed to now) then the newly
created row will not be removed.

So far as I can determine in testing, this works perfectly and provides
a stable inventory screen, regardless of the number of refineries, etc.
present.