Retain voxel shape information while floating ore exists.
MyFloatingObject entities still collide with voxel objects when grids and
players are out of range. If shape information has been discarded this
will cause expensive RequestShapeBlocking calls to happen frequently on the
simulation thread, with a noticeable effect on sim speed.

This fix tracks floating objects too and retains shape information as long
as they're in range.

refs issue #376