Remove TryDoZipAction and use SplitZipFilePath instead.
Splitting the path only once in GetFiles is a tiny performance win since
File.Exists gets called one fewer times, although the main goal here is to
separate two distinct use cases for TryDoZipAction: opening a zip stream,
and checking for existence of files. The latter is a candidate for
caching.