Assembler Fix
Assemblers checked if they had 540 kW of power, not if they have enough
power to run taking the energy efficiency modules into account.

I'm not sure if you want to use the GetOperationalPowerConsumption
method you have in MyAssembler or run the check somewhere else, but this
should fix the bug reported in this thread:

http://forum.keenswh.com/threads/01-097-008-not-enough-power-assembler.7366828/

For my work I have to stick with VS2012, so I can't build Space
Engineers, so this is not tested.