Adds a new type of turret, LargeGenericTurret. It's a simple interior style turret, but with support for conveyor networks. Perfect for turrets that aren't gatling guns or rocket launchers, such as autocannons, tank guns and such.
Internally, the models use the same structure as the interior turrets, but named differently. The subpart for the turret body is GenericTurretBase1 and the subpart for the barrel(s) of the turret is GenericTurretBase2. Everything else behaves as a Space Engineers turret behaves. The main change was setting the base class from LargeTurretBase to LargeConveyorTurretBase. It's been tested and works.

Signed-off-by: Martyn Mees <cmdbob@gmail.com>