Exposed AirDensity around an IMyAirVent
Using this extension, the AirDensity of the closest planet can be
determined with any IMyAirVent.

The "gameplay explanation" behind this is that every air vent can
measure oxygen already (though only for rooms) so it has to be able to
measure the general density of the surrounding atmosphere as well.

Exposed value:
float IMyAirVent.GetAirDensity()

Ingame script example:
http://pastebin.com/AQMyBRyN