Fix projectors not working in offline mode
I'm not 100% sure this is the correct fix, however the current functionality ignores players in offline mode as admins as m_adminMode does not contain the user. MySession.Static.IsCopyPastingEnabled thus returns false.

Due to the fact that projectors specifically check for if IsCopyPastingEnabled the false flag stops any blueprints in offline mode from working.